+++ src/RtpParameters.ts
@@ -99,7 +99,7 @@ export type RtpHeaderExtension =
 	 * Media kind. If empty string, it's valid for all kinds.
 	 * Default any media kind.
 	 */
-	kind?: MediaKind | '';
+	kind: MediaKind;
 
 	/*
 	 * The URI of the RTP header extension, as defined in RFC 5285.