+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ### Unreleased
-* `kind` field of `RtpHeaderExtension` is no longer optional
+* `kind` field of `RtpHeaderExtension` is no longer optional.
+* Refactor API inconsistency in internal RTP Observer communication with worker.
 
 ### 3.7.12