+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-### Unreleased
 
-* `kind` field of `RtpHeaderExtension` is no longer optional.
+### 3.7.15
+
+* `kind` field of `RtpHeaderExtension` is no longer optional. It must be 'audio' or 'video'.
 * Refactor API inconsistency in internal RTP Observer communication with worker.
+
 
 ### 3.7.14